requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
gspread>=6.1.0
google-auth>=2.33.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
requests>=2.32.3
aiohttp>=3.9.5
